requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
transformers==4.23.1
